add pipeline object to the library initialize it with a right-reference
